feat(client|redux): general mocks - returns
