--- packages/client/src/entities/schemas/returnItem.js
@@ -1,9 +1,0 @@
-import { schema } from 'normalizr';
-
-export default new schema.Entity(
-  'returnItems',
-  {},
-  {
-    idAttribute: 'orderItemId',
-  },
-);
