--- packages/redux/src/returns/__fixtures__/returns.fixtures.js
@@ -1,464 +1,0 @@
-/**
- * Response payloads.
- */
-
-export const responses = {
-  post: {
-    success: {
-      availableDates: [
-        '/Date(1549538022607+0000)/',
-        '/Date(1549624422607+0000)/',
-        '/Date(1549883622607+0000)/',
-        '/Date(1549970022607+0000)/',
-        '/Date(1550056422607+0000)/',
-        '/Date(1550142822607+0000)/',
-        '/Date(1550229222607+0000)/',
-        '/Date(1550488422607+0000)/',
-        '/Date(1550574822607+0000)/',
-        '/Date(1550661222607+0000)/',
-        '/Date(1550747622607+0000)/',
-      ],
-      awbUrl: '/pt/api/returns/5926969/awb',
-      baseUrl: null,
-      clientOnlyPage: false,
-      components: {},
-      countryCode: null,
-      countryId: 0,
-      courier: 1,
-      createdDate: '/Date(1549538021917)/',
-      cultureCode: null,
-      currencyCode: null,
-      currencyCultureCode: null,
-      dataLayer: null,
-      id: 5926969,
-      invoiceUrl: '/pt/api/returns/5926969/invoice',
-      isMobileDevice: false,
-      items: [
-        {
-          id: 0,
-          orderItemId: 10076742,
-          reason: 'Item does not fit',
-          description: '',
-          status: 0,
-        },
-      ],
-      maximumDateForPickup: '/Date(1550793599000)/',
-      merchantId: 10973,
-      merchantLocationId: null,
-      newsletterSubscriptionOptionDefault: false,
-      numberOfBoxes: 1,
-      numberOfItems: 1,
-      orderId: '8VXRHN',
-      pageContent: null,
-      pageType: null,
-      pickupSchedule: {
-        start: '/Date(1549965600000)/',
-        end: '/Date(1549994400000)/',
-      },
-      redirectUrl: null,
-      references: [],
-      relativeUrl: null,
-      requestSourceCountryCode: null,
-      returnUrl: null,
-      screenPixelsHeight: 0,
-      screenPixelsWidth: 0,
-      seoMetadata: null,
-      seoPageType: 1,
-      serverSideJsApp: null,
-      siteKeys: {},
-      slug: null,
-      staticPath: null,
-      status: 2,
-      subfolder: null,
-      translationsUrl: null,
-      type: 0,
-      userId: 29511627,
-      userPickupAddress: {
-        addressLine1: 'Rua da Lionesa nº 446, Edificio G1',
-        addressLine2: '',
-        addressLine3: null,
-        city: {
-          countryId: 165,
-          id: 0,
-          name: 'Leca do Balio',
-          stateId: null,
-        },
-        country: {
-          alpha2Code: 'PT',
-          alpha3Code: 'PRT',
-          continentId: 3,
-          culture: 'pt-PT',
-          id: 165,
-          name: 'Portugal',
-          nativeName: 'Portugal',
-          region: 'Europe',
-          regionId: 0,
-          subRegion: null,
-          subfolder: null,
-        },
-        ddd: null,
-        firstName: 'Rafael',
-        id: '00000000-0000-0000-0000-000000000000',
-        lastName: null,
-        neighbourhood: null,
-        phone: '4564132',
-        state: null,
-        vatNumber: null,
-        zipCode: '4465-671',
-      },
-    },
-  },
-  get: {
-    success: {
-      availableDates: [
-        '/Date(1549538022607+0000)/',
-        '/Date(1549624422607+0000)/',
-        '/Date(1549883622607+0000)/',
-        '/Date(1549970022607+0000)/',
-        '/Date(1550056422607+0000)/',
-        '/Date(1550142822607+0000)/',
-        '/Date(1550229222607+0000)/',
-        '/Date(1550488422607+0000)/',
-        '/Date(1550574822607+0000)/',
-        '/Date(1550661222607+0000)/',
-        '/Date(1550747622607+0000)/',
-      ],
-      awbUrl: '/pt/api/returns/5926969/awb',
-      baseUrl: null,
-      clientOnlyPage: false,
-      components: {},
-      countryCode: null,
-      countryId: 0,
-      courier: 1,
-      createdDate: '/Date(1549538021917)/',
-      cultureCode: null,
-      currencyCode: null,
-      currencyCultureCode: null,
-      dataLayer: null,
-      id: 5926969,
-      invoiceUrl: '/pt/api/returns/5926969/invoice',
-      isMobileDevice: false,
-      items: [
-        {
-          id: 0,
-          orderItemId: 10076742,
-          reason: 'Item does not fit',
-          description: '',
-          status: 0,
-        },
-      ],
-      maximumDateForPickup: '/Date(1550793599000)/',
-      merchantId: 10973,
-      merchantLocationId: null,
-      newsletterSubscriptionOptionDefault: false,
-      numberOfBoxes: 1,
-      numberOfItems: 1,
-      orderId: '8VXRHN',
-      pageContent: null,
-      pageType: null,
-      pickupSchedule: {
-        start: '/Date(1549965600000)/',
-        end: '/Date(1549994400000)/',
-      },
-      redirectUrl: null,
-      references: [],
-      relativeUrl: null,
-      requestSourceCountryCode: null,
-      returnUrl: null,
-      screenPixelsHeight: 0,
-      screenPixelsWidth: 0,
-      seoMetadata: null,
-      seoPageType: 1,
-      serverSideJsApp: null,
-      siteKeys: {},
-      slug: null,
-      staticPath: null,
-      status: 2,
-      subfolder: null,
-      translationsUrl: null,
-      type: 0,
-      userId: 29511627,
-      userPickupAddress: {
-        addressLine1: 'Rua da Lionesa nº 446, Edificio G1',
-        addressLine2: '',
-        addressLine3: null,
-        city: {
-          countryId: 165,
-          id: 0,
-          name: 'Leca do Balio',
-          stateId: null,
-        },
-        country: {
-          alpha2Code: 'PT',
-          alpha3Code: 'PRT',
-          continentId: 3,
-          culture: 'pt-PT',
-          id: 165,
-          name: 'Portugal',
-          nativeName: 'Portugal',
-          region: 'Europe',
-          regionId: 0,
-          subRegion: null,
-          subfolder: null,
-        },
-        ddd: null,
-        firstName: 'Rafael',
-        id: '00000000-0000-0000-0000-000000000000',
-        lastName: null,
-        neighbourhood: null,
-        phone: '4564132',
-        state: null,
-        vatNumber: null,
-        zipCode: '4465-671',
-      },
-    },
-  },
-  getPickupCapabilities: {
-    success: {
-      availableEndHours: [
-        '/Date(1549987200000)/',
-        '/Date(1549990800000)/',
-        '/Date(1549994400000)/',
-        '/Date(1549998000000)/',
-        '/Date(1550001600000)/',
-        '/Date(1550005200000)/',
-        '/Date(1550008800000)/',
-        '/Date(1550012400000)/',
-      ],
-      availableStartHours: [
-        '/Date(1549962000000)/',
-        '/Date(1549965600000)/',
-        '/Date(1549969200000)/',
-        '/Date(1549972800000)/',
-        '/Date(1549976400000)/',
-        '/Date(1549980000000)/',
-        '/Date(1549983600000)/',
-        '/Date(1549987200000)/',
-      ],
-      availableTimeSlots: [
-        {
-          start: '/Date(1549987200000)/',
-          end: '/Date(1549962000000)/',
-        },
-        {
-          start: '/Date(1549987200000)/',
-          end: '/Date(1549965600000)/',
-        },
-        {
-          start: '/Date(1549987200000)/',
-          end: '/Date(1549969200000)/',
-        },
-        {
-          start: '/Date(1549987200000)/',
-          end: '/Date(1549972800000)/',
-        },
-        {
-          start: '/Date(1549987200000)/',
-          end: '/Date(1549976400000)/',
-        },
-      ],
-      pickupDate: '/Date(1549929600000)/',
-    },
-  },
-  getReturnsFromOrder: {
-    get: {
-      success: [
-        {
-          availableDates: [
-            '/Date(1549538022607+0000)/',
-            '/Date(1549624422607+0000)/',
-            '/Date(1549883622607+0000)/',
-            '/Date(1549970022607+0000)/',
-            '/Date(1550056422607+0000)/',
-            '/Date(1550142822607+0000)/',
-            '/Date(1550229222607+0000)/',
-            '/Date(1550488422607+0000)/',
-            '/Date(1550574822607+0000)/',
-            '/Date(1550661222607+0000)/',
-            '/Date(1550747622607+0000)/',
-          ],
-          awbUrl: '/pt/api/returns/5926969/awb',
-          baseUrl: null,
-          clientOnlyPage: false,
-          components: {},
-          countryCode: null,
-          countryId: 0,
-          courier: 1,
-          createdDate: '/Date(1549538021917)/',
-          cultureCode: null,
-          currencyCode: null,
-          currencyCultureCode: null,
-          dataLayer: null,
-          id: 5926969,
-          invoiceUrl: '/pt/api/returns/5926969/invoice',
-          isMobileDevice: false,
-          items: [
-            {
-              id: 0,
-              orderItemId: 10076742,
-              reason: 'Item does not fit',
-              description: '',
-              status: 0,
-            },
-          ],
-          maximumDateForPickup: '/Date(1550793599000)/',
-          merchantId: 10973,
-          merchantLocationId: null,
-          newsletterSubscriptionOptionDefault: false,
-          numberOfBoxes: 1,
-          numberOfItems: 1,
-          orderId: '8VXRHN',
-          pageContent: null,
-          pageType: null,
-          pickupSchedule: {
-            start: '/Date(1549965600000)/',
-            end: '/Date(1549994400000)/',
-          },
-          redirectUrl: null,
-          references: [
-            {
-              name: 3,
-              url: '/en-pt/api/returns/5926969/references/ReturnLabelAWB',
-            },
-          ],
-          relativeUrl: null,
-          requestSourceCountryCode: null,
-          returnUrl: null,
-          screenPixelsHeight: 0,
-          screenPixelsWidth: 0,
-          seoMetadata: null,
-          seoPageType: 1,
-          serverSideJsApp: null,
-          siteKeys: {},
-          slug: null,
-          staticPath: null,
-          status: 2,
-          subfolder: null,
-          translationsUrl: null,
-          type: 0,
-          userId: 29511627,
-          userPickupAddress: {
-            addressLine1: 'Rua da Lionesa nº 446, Edificio G1',
-            addressLine2: '',
-            addressLine3: null,
-            city: {
-              countryId: 165,
-              id: 0,
-              name: 'Leca do Balio',
-              stateId: null,
-            },
-            country: {
-              alpha2Code: 'PT',
-              alpha3Code: 'PRT',
-              continentId: 3,
-              culture: 'pt-PT',
-              id: 165,
-              name: 'Portugal',
-              nativeName: 'Portugal',
-              region: 'Europe',
-              regionId: 0,
-              subRegion: null,
-              subfolder: null,
-            },
-            ddd: null,
-            firstName: 'Rafael',
-            id: '00000000-0000-0000-0000-000000000000',
-            lastName: null,
-            neighbourhood: null,
-            phone: '4564132',
-            state: null,
-            vatNumber: null,
-            zipCode: '4465-671',
-          },
-        },
-      ],
-    },
-  },
-  patch: {
-    success: {
-      redirectUrl:
-        'https://api.blackout.com/pt/account/return/summary/?id=5926969',
-    },
-  },
-};
-
-export const orderId = responses.get.success.orderId;
-export const returnId = responses.get.success.id;
-export const returnItemId = responses.get.success.items[0].id;
-export const orderItemId = responses.get.success.items[0].orderItemId;
-export const returnsNormalizedPayload = {
-  entities: {
-    returnItems: {
-      [orderItemId]: {
-        id: 0,
-        orderItemId,
-        reason: 'Item does not fit',
-        description: '',
-        status: 0,
-      },
-    },
-    returns: {
-      [returnId]: {
-        id: returnId,
-        orderId: '8VXRHN',
-        merchantId: 10973,
-        userId: 29511627,
-        type: 0,
-        status: 2,
-        courier: 1,
-        numberOfBoxes: 1,
-        numberOfItems: 1,
-        items: [orderItemId],
-        createdDate: 1549538021917,
-        availableDates: [
-          1549538022607, 1549624422607, 1549883622607, 1549970022607,
-          1550056422607, 1550142822607, 1550229222607, 1550488422607,
-          1550574822607, 1550661222607, 1550747622607,
-        ],
-        userPickupAddress: {
-          addressLine1: 'Rua da Lionesa nº 446, Edificio G1',
-          addressLine2: '',
-          addressLine3: null,
-          city: {
-            countryId: 165,
-            id: 0,
-            name: 'Leca do Balio',
-            stateId: null,
-          },
-          country: {
-            alpha2Code: 'PT',
-            alpha3Code: 'PRT',
-            continentId: 3,
-            culture: 'pt-PT',
-            id: 165,
-            name: 'Portugal',
-            nativeName: 'Portugal',
-            region: 'Europe',
-            regionId: 0,
-            subRegion: null,
-            subfolder: null,
-          },
-          ddd: null,
-          firstName: 'Rafael',
-          id: '00000000-0000-0000-0000-000000000000',
-          lastName: null,
-          neighbourhood: null,
-          phone: '4564132',
-          state: null,
-          vatNumber: null,
-          zipCode: '4465-671',
-        },
-        maximumDateForPickup: 1550793599000,
-        pickupSchedule: {
-          start: 1549965600000,
-          end: 1549994400000,
-        },
-      },
-    },
-  },
-  result: returnId,
-};
-export const returnsFromOrderNormalizedPayload = {
-  ...returnsNormalizedPayload,
-  result: [returnId],
-};
